Resolving style issues
