+++ tests/test_server.py
@@ -1,31 +1,6 @@
 import json
-from pytest import fixture
 
-from get_drunk_telegram_bot.bot.server import create_server
-
-from utils import TelegramInterfaceMocker
-
-
-class FakeArgs:
-    token = None
-    web_hook_url = None
-    debug = False
-
-
-@fixture
-def client():
-    with TelegramInterfaceMocker():
-        test_app = create_server(FakeArgs())
-        yield test_app.test_client()
-
-
-def make_message_from_text(text):
-    return {
-        'message': {
-            'chat': {'id': ""},
-            'text': text
-        }
-    }
+from utils import make_message_from_text, client
 
 
 def test_server_responses(client):
