+++ tests/utils.py
@@ -1,6 +1,7 @@
 from unittest.mock import patch
+from pytest import fixture
 
-from get_drunk_telegram_bot.bot.server import GetDrunkBotHandler
+from get_drunk_telegram_bot.bot.server import GetDrunkBotHandler, create_server
 
 
 class FakeInterface:
@@ -32,3 +33,56 @@ class TelegramInterfaceMocker:
 
     def __exit__(self, *args, **kwargs):
         return self.patcher.__exit__(*args, **kwargs)
+
+
+def canonical_normalization(text):
+    return '\n'.join([line.strip() for line in text.split('\n')])
+
+
+@fixture
+def tmp_dir(tmpdir):
+    return str(tmpdir)
+
+
+@fixture
+def client():
+    with TelegramInterfaceMocker():
+        test_app = create_server(FakeArgs())
+        yield test_app.test_client()
+
+
+class FakeArgs:
+    token = None
+    web_hook_url = None
+    debug = False
+
+
+def make_message_from_text(text):
+    return {
+        'message': {
+            'chat': {'id': ""},
+            'text': text
+        }
+    }
+
+
+def get_handler():
+    with TelegramInterfaceMocker():
+        return GetDrunkBotHandler()
+
+
+def run_test_request(handler, text, control_normalization=True):
+    with TelegramInterfaceMocker():
+        handler.process_message("", text)
+
+        text = getattr(handler, 'saved_text', None)
+        im_path = getattr(handler, 'saved_photo_path', None)
+
+        # Maybe some assert here (common for all requests)
+        if control_normalization:
+            assert canonical_normalization(text) == text, (
+                "No indents in response."
+                "Use GetDrunkBotHandler.normalize_text(text)."
+            )
+
+        return (text, im_path)
